+++ plugin/aiusageplugin.cpp
@@ -10,6 +10,7 @@
 #include "deepseekprovider.h"
 #include "groqprovider.h"
 #include "xaiprovider.h"
+#include "ollamacloudprovider.h"
 #include "openrouterprovider.h"
 #include "togetherprovider.h"
 #include "cohereprovider.h"
@@ -46,6 +47,7 @@ void AiUsagePlugin::registerTypes(const char *uri)
     qmlRegisterType<DeepSeekProvider>(uri, 1, 0, "DeepSeekProvider");
     qmlRegisterType<GroqProvider>(uri, 1, 0, "GroqProvider");
     qmlRegisterType<XAIProvider>(uri, 1, 0, "XAIProvider");
+    qmlRegisterType<OllamaCloudProvider>(uri, 1, 0, "OllamaCloudProvider");
     qmlRegisterType<OpenRouterProvider>(uri, 1, 0, "OpenRouterProvider");
     qmlRegisterType<TogetherProvider>(uri, 1, 0, "TogetherProvider");
     qmlRegisterType<CohereProvider>(uri, 1, 0, "CohereProvider");